+++ src/dom_elements.js
@@ -1,0 +1,6 @@
+const scoresList = document.getElementsByClassName('scores-list')[0]
+
+
+export {
+  scoresList,
+}
