+++ src/get_scores.js
@@ -1,0 +1,10 @@
+
+const scores = [
+  {name: 'Name1', score: '262'},
+  {name: 'Name3', score: '252'},
+  {name: 'Name2', score: '22'},
+  {name: 'Name4', score: '242'},
+  {name: 'Name', score: '212'},
+]
+
+export { scores };
