+++ src/index.html
@@ -20,7 +20,5 @@
       <button id="submit">Submit</button>
     </div>
   </main>
-  <!-- <h3>Add your score</h3> -->
-
 </body>
 </html>
